fix(sheet): clear style util
